lashd: Make Progress signal sending behavior more consistent.
	- When a task commences always send Progress signal with value 0.
	- When a task finishes always send Progress signal with value 100.
	- Never send Progress signal of 0 or 100 during a task's lifetime.

In the past these signals weren't behaving consistently for Save and Load tasks.
Also, a Progress signal with value 0 signified both the beginning and end of a
task, leaving it up to the controller application to decipher which is which.
This is now fixed and apps can rely on 0 meaning 'go' and 100 meaning 'done'.