waf: Disable opimizations in debug mode
The -Wuninitilized vs optimizations situation is problematic.
-Wuninitialized needs optimizations to work properly
but then binaries are hard to debug because stack gets optimized.

The implemented solution is not to build debuggable executables
by default, even for non-release source trees. If debug mode is
force enabled at configure stage, then optimizations are force
disabled. This way use of uninitialized data will be catched
early during development and when debugging is needed, the debug
mode will have to be force enabled.