Revert "Fixed uninitialized vars in rooms created from templates"
This reverts commit 9bf795905f4ac22d1add6166ac75f13cb1c0d7d4.

ladish_room_create_interal() is suposed to allocate the struct and
initialize the name and uuid members. The template member is supposed to
be initialized by the caller functions, either ladish_room_create() or
ladish_room_create_template(). The siblings member is initialized
later, when the room is actually added to the appropriate list.

The graph member is used by both room templates and room objects. This
will be fixed in the following commit.

All other ladish_room members are not valid for room template objects.