Implement JS delay workaround. #187
yoshimi-0.060.10 jack session implementation is broken. Save complete
reply is sent from the jack session callback but the actual save is
made asynchronously, in the main thread, on next main loop iteration.

This changeset implements workaround for app bugs like this.
User is able to set JS delay in seconds. Default is zero.