meson: include build directory no matter it's name
when updating the ebuild i noticed that the include for the config.h and
version.h files relied on the build directory being name "build". This
is fixed now by using the last segment of the meson.build_root() output
which is the correct build dir.

also delete no longer needed manual creation of version.h and config.h